Comment out support for msedgedriver on Linux until it's out of preview.
The linux driver version is in preview and not in sync with other operating systems. This means that the LATEST_RELEASE_ file is not usable for it. It has a separate _LINUX file AND the download file lives in a deferent location than the current `base_url`. This combination of differences makes it difficult to support Linux builds within the current implementation. I was able to make it sort of work, but the implementation was messy and seemed hack-ish. I think the best thing to do is to hold off on supporting Linux builds until the general release is available. It's currently in a preview release (insider) only.